feat: add roll modifiers for wounded state (#793)
This release adds automatic roll modifiers for the wounded state when using lifeblood and stamina as the damage characteristics.  Theses roll modifiers can be customized in the ruleset settings.  The feature is useful for systems such as Cepheus Deluxe.  Also, lifeblood and stamina damage now constrained to be >=0 and not more than the base value.