feat: Include User Macros in Compendium (#981)
This update introduces a big shift in the way we maintain macros related to Twodsix.  Going forward, macros (formerly in the user macro folder on the wiki) and their related tables will be in compendiums that are part of the system release.  We won't maintain the macros on the wiki any longer (and will likely delete them at some point).  All the relevant macros from the wiki are included with this release.  A few have been updated to be in sync with the current code base.  This should make it much easier for folks to use and find the macros.