feat: add vehicle sheet (#1015)
With this release we finally have a vehicle sheet!  There is an option under the ruleset settings to select what is shown in the protection box: armor only, damage thresholds, and armor/hull/structure.  Most of the systems have a default set.  Those playing MgT2 should use Armor Only as that is a freeform text box to create what you want.