feat: add initiative roll to actor sheet (#1043)
This release adds an optional button to roll initiative from actor sheet (check box in display settings). The button appears at the top of the skills tab.  If the actor is not present in combat tracker, the actor/token will be automatically added.  If the actor already has an initiative value, a warning appears instead and no roll is made.  The actor must be present in the scene, as well.